requirements.txt
blinker==1.6.2
cachelib==0.13.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==43.0.0
flask==2.3.3
flask-session==0.8.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
identity==0.5.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.30.0
msgspec==0.18.6
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.16
werkzeug==2.3.7
zipp==3.16.2
